+++ square_file_store_helper/pydantic_models.py
@@ -1,0 +1,11 @@
+from typing import List
+
+from pydantic import BaseModel
+
+
+class UploadFileV0Response(BaseModel):
+    main: str
+
+
+class DeleteFilesV0Response(BaseModel):
+    main: List[str]